requirements.txt
requests>=2.28.0
pandas>=1.5.0

# Google Sheets integration
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Configuration management
python-dotenv>=0.20.0

# Resilience & retry logic
tenacity>=8.2.0

# Development & Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=22.12.0
flake8>=5.0.0
mypy>=0.990
# Core dependencies
requests>=2.28.0
pandas>=1.5.0

# Google Sheets integration
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Configuration management
python-dotenv>=0.20.0

# Resilience & retry logic
tenacity>=8.2.0

# Development & Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=22.12.0
flake8>=5.0.0
mypy>=0.990